# specs/008-agent-construction/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Agent API
  description: API for the RAG-powered AI assistant that answers questions from the Physical AI & Humanoid Robotics book
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/agent/query:
    post:
      summary: Submit a query to the RAG agent
      description: Sends a user query to the AI agent which retrieves relevant book content and generates a grounded response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIRequest'
      responses:
        '200':
          description: Successful response from the agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the agent service
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
components:
  schemas:
    APIRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's query text
          example: "What are the key components of humanoid robotics?"
        user_context:
          type: object
          description: Context about the user or session
          example: {}
        retrieval_params:
          type: object
          description: Parameters for the retrieval process
          example: {"top_k": 5, "threshold": 0.7}
        grounding_required:
          type: boolean
          description: Whether strict grounding is required
          default: true
    APIResponse:
      type: object
      required:
        - success
        - answer
        - sources
        - timestamp
        - query_id
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        answer:
          type: string
          description: The agent's answer to the query
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          description: List of sources used to generate the answer
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
        query_id:
          type: string
          description: ID of the original query
        error:
          type: string
          description: Error message if the request failed
    Source:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the content chunk
        content:
          type: string
          description: The actual text content from the book
        source_url:
          type: string
          description: URL or reference to the original source
        page_title:
          type: string
          description: Title of the page/chapter where content appears
        similarity_score:
          type: number
          description: How similar this content is to the query
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          description: Always false for error responses
        error:
          type: string
          description: Description of the error that occurred
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          description: Health status of the service
          enum: [healthy, degraded, unavailable]
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: When the health check was performed
        details:
          type: object
          description: Additional health check details
          example: {"database": "connected", "vector_db": "available"}